+++ DA3.Common/ViewModel/CartViewModel.cs
@@ -2,11 +2,11 @@ namespace DA3.Models
 {
     public class CartViewModel
     {
-        public decimal PriceShipFree { get; set; }
+        public double PriceShipFree { get; set; }
 
-        public decimal PriceShipStandard { get; set; }
+        public double PriceShipStandard { get; set; }
 
-        public decimal PriceShipExpress { get; set; }
+        public double PriceShipExpress { get; set; }
 
         public CartModel CartModel { get; set; } = new CartModel();
     }
