+++ DA3.DAL/Domain/OrderDetails.cs
@@ -10,7 +10,7 @@
 
         public virtual string? ProductId { get; set; }
 
-        public virtual decimal Price { get; set; }
+        public virtual double Price { get; set; }
 
         public virtual int Quantity { get; set; }
     }
